I think it may just be done

+++ editor-backend/app.py
@@ -6,11 +6,13 @@ import json
 import logging
 import os
 import pathlib
+import shlex
 import shutil
 import subprocess
 import typing
 
-import pydantic
+import betterproto
+import proto
 from starlette.applications import Starlette
 from starlette.endpoints import HTTPEndpoint, WebSocketEndpoint
 from starlette.middleware import Middleware
@@ -28,6 +30,10 @@ file = pathlib.Path(__file__)
 PSEUDO_WORKSPACE_PATH = file.parent.parent / "pseudo"
 COMPILER_PATH = PSEUDO_WORKSPACE_PATH / "target" / "debug" / "pseudo-cli"
 LIB_PATH = PSEUDO_WORKSPACE_PATH / "target" / "debug" / "libpseudo_sys.so"
+
+CPU_TIME_LIMIT = 1
+MEMORY_LIMIT = 1024*1024*16
+SOURCE_LIMIT = 1024*64
 
 # make sure the compiler is built
 subprocess.run([
@@ -74,28 +80,17 @@ job_ids = itertools.count()
 jobs: dict[int, Job] = dict()
 
 
-class JobResponseSuccess(pydantic.BaseModel):
-    type: typing.Literal["success"]
-    job_id: int
-
-
-class JobResponseFailure(pydantic.BaseModel):
-    type: typing.Literal["failure"]
-    error: typing.Any
-
-
-class JobResponse(pydantic.BaseModel):
-    message: typing.Union[JobResponseSuccess, JobResponseFailure] = pydantic.Field(..., discriminator="type")
-
-
 class JobEndpoint(HTTPEndpoint):
-    class Post(pydantic.BaseModel):
-        code: str
-        language_enable_lists: bool
-
     async def post(self, request: Request):
-        body = await request.body()
-        parsed_request: self.Post = self.Post.parse_raw(body)
+        body = await request.json()
+        parsed_request: proto.PostJob = proto.PostJob().from_dict(body)
+
+        if len(parsed_request.code) > SOURCE_LIMIT:
+            response = proto.JobResponse(
+                other_error="Eroare: codul sursă depășește dimensiunea maximă."
+            )
+
+            return JSONResponse(response.to_dict())
 
         job_id = next(job_ids)
         job = Job(job_id, "created")
@@ -127,33 +122,31 @@ class JobEndpoint(HTTPEndpoint):
                 error = None
                 logger.info("failed to decode compiler error %s", stdout)
 
-            return JSONResponse(dict(message=dict(type="failure", error=error)))
+            response = proto.JobResponse(
+                parser_error=proto.JobResponseParserError(
+                    line=error["line"],
+                    column=error["column"],
+                    message=error["message"],
+                )
+            )
+
+            return JSONResponse(response.to_dict())
         job.status = "compiled"
+
+        response = proto.JobResponse(
+            success=proto.JobResponseSuccess(
+                job_id=job_id,
+            )
+        )
         
-        return JSONResponse(dict(message=dict(type="success", job_id=job_id)))
-
-
-class ReceiveStdin(pydantic.BaseModel):
-    type: typing.Literal["stdin"]
-    stdin: str
-
-
-class ReceiveStop(pydantic.BaseModel):
-    type: typing.Literal["stop"]
-
-
-class Receive(pydantic.BaseModel):
-    message: typing.Union[ReceiveStdin, ReceiveStop] = pydantic.Field(..., discriminator="type")
+        return JSONResponse(response.to_dict())
 
 
 class JobStatusEndpoint(WebSocketEndpoint):
     encoding = "json"
 
-    class Send(pydantic.BaseModel):
-        stdout: str
-    
     @staticmethod
-    async def _send_stdout(
+    async def _monitor_process(
         job_id: int,
         websocket: WebSocket,
         process: asyncio.subprocess.Process,
@@ -163,27 +156,78 @@ class JobStatusEndpoint(WebSocketEndpoint):
             data = await process.stdout.read(256)
             final = len(data) == 0
             decoded = decoder.decode(data, final)
-            await websocket.send_json(dict(JobStatusEndpoint.Send(stdout=decoded)))
+
+            message = proto.JobOutput(
+                stdout=proto.JobOutputStdout(
+                    text=decoded,
+                )
+            )
+            
+            await websocket.send_json(message.to_dict())
             if final:
                 break
-        await websocket.close()
 
-        job = jobs[job_id]
-        del jobs[job_id]
-        shutil.rmtree(job.state_path)
+        await process.communicate()
+
+        if process.returncode != 0:
+            logger.info("non-zero return code %s", process.returncode)
+
+        match process.returncode:
+            case 0:
+                pass
+            
+            # NOTE: Return code of 139 means that the process was ended with
+            # SIGKILL. Since we're running the process using prlimit, we're going to
+            # assume that it was killed because it exceeded its allotted limits.
+            case 139:
+                message = proto.JobOutput(
+                    stdout=proto.JobOutputStdout(
+                        text="Eroare: programul a depășit resursele de procesor care i-au fost alocate.\n",
+                    )
+                )
+            
+                await websocket.send_json(message.to_dict())
+
+            # Return code of 134 _typically_ means segfault. And return code 137
+            # means out of memory.
+            case item if item in [134, 137]:
+                message = proto.JobOutput(
+                    stdout=proto.JobOutputStdout(
+                        text="Eroare: programul a depășit memoria care i-a fost alocată.\n",
+                    )
+                )
+            
+                await websocket.send_json(message.to_dict())
+
+            case other:
+                logger.warning("unexpected return code %s", other)
+
+        await websocket.close()
+        if (job := jobs.get(job_id)) is not None:
+            del jobs[job_id]
+            shutil.rmtree(job.state_path)
 
     async def on_connect(self, websocket: WebSocket):
         job_id = int(websocket.path_params["job_id"])
         job = jobs[job_id]
 
-        # NOTE: Use script command to run the executable, so that its stdout is
-        # line-buffered. Many thanks to this[0] stackexchange post.
-        # [0]: https://unix.stackexchange.com/a/61833
         process = await asyncio.subprocess.create_subprocess_exec(
+            # NOTE: Use script command to run the executable, so that its stdout
+            # is line-buffered. Many thanks to this[0] stackexchange post.
+            # [0]: https://unix.stackexchange.com/a/61833
             "script",
             "-q",
-            "-c", str(job.executable_path),
             "-E", "never",
+            "-c", shlex.join([
+                # NOTE: Use prlimit command, to limit the CPU time that the
+                # executable is allowed to run for.
+                "prlimit",
+                f"--cpu={CPU_TIME_LIMIT}",
+                f"--data={MEMORY_LIMIT}",
+                "--",
+                str(job.executable_path),
+            ]),
+            "-e",
             "/dev/null",
             stdout=asyncio.subprocess.PIPE,
             stdin=asyncio.subprocess.PIPE,
@@ -191,7 +235,7 @@ class JobStatusEndpoint(WebSocketEndpoint):
         job.status = "running"
         job.process = process
 
-        asyncio.create_task(self._send_stdout(job_id, websocket, process))
+        asyncio.create_task(self._monitor_process(job_id, websocket, process))
         
         await websocket.accept()
 
@@ -199,11 +243,11 @@ class JobStatusEndpoint(WebSocketEndpoint):
         job_id = int(websocket.path_params["job_id"])
         job = jobs[job_id]
 
-        parsed_data: Receive = Receive.parse_obj(data)
-        logger.info("job %s %s", job_id, parsed_data.message.type)
-        match parsed_data.message.type:
+        message: proto.JobInput = proto.JobInput().from_dict(data)
+        type_, message = betterproto.which_one_of(message, "message")
+        match type_:
             case "stdin":
-                job.process.stdin.write(parsed_data.message.stdin.encode())
+                job.process.stdin.write(message.text.encode())
                 await job.process.stdin.drain()
             case "stop":
                 job.process.kill()
